requirements.txt
aioboto3==11.1.0
aiobotocore==2.5.0
aiofauna==0.0.11
aiohttp==3.8.4
aiohttp-sse==2.1.0
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.76
botocore==1.29.76
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
orjson==3.8.10
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
sse-starlette==1.3.3
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
yarl==1.8.2
